Just practicing, main work is blackjack
Since McGill does not value actual coding and hasn’t assigned any real
coding projects in the last year, I decided to practice a bit on my
own. I started with making a prime number calculator and then made a
file to reverse strings. This helped me brush up on arithmetic and
string manipulation in java, which are both fairly simple. Then I made
blackjack as a bit more of a challenge. It was still a fairly simple
project, but required a bit more coding thinking with for loops and
slightly more complicated if statements.